requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
# psycopg2-binary==2.9.9 # For production PostgreSQL
# celery==5.3.6          # For production background tasks
# redis==5.0.1           # For production cache
openai==1.14.0
scikit-learn==1.3.2
pandas==2.1.4
argon2-cffi==23.1.0
